refactor: remove MigrationView/Footer v1 component
