+++ operate/client/src/App/Processes/MigrationView/index.tsx
@@ -13,12 +13,10 @@ import {processInstanceMigrationStore} from 'modules/stores/processInstanceMigra
 import {processesStore} from 'modules/stores/processes/processes.migration';
 import {TopPanel} from './TopPanel';
 import {BottomPanel} from './BottomPanel';
-import {Footer} from './Footer';
-import {Footer as FooterV2} from './Footer/v2';
+import {Footer} from './Footer/v2';
 import {PAGE_TITLE} from 'modules/constants';
 import {MigrationSummaryNotification} from './MigrationSummaryNotification';
 import {observer} from 'mobx-react';
-import {IS_BATCH_OPERATIONS_V2} from 'modules/feature-flags';
 
 const MigrationView: React.FC = observer(() => {
   useEffect(() => {
@@ -49,7 +47,7 @@ const MigrationView: React.FC = observer(() => {
         type="migrate"
         topPanel={<TopPanel />}
         bottomPanel={<BottomPanel />}
-        footer={IS_BATCH_OPERATIONS_V2 ? <FooterV2 /> : <Footer />}
+        footer={<Footer />}
         additionalTopContent={
           processInstanceMigrationStore.isSummaryStep ? (
             <MigrationSummaryNotification />
